data class rearange
