+++ app/src/main/java/com/example/tabs/utils/models/Contact.kt
@@ -1,0 +1,12 @@
+package com.example.tabs.utils.models
+
+import java.util.Date
+import com.example.tabs.utils.models.PresentHistory
+
+data class Contact(
+    val name: String,
+    val phoneNumber: String,
+    val bDay: Date,
+    val recentContact: Int,
+    val presentHistory: List<PresentHistory>
+)
