+++ app/src/main/java/com/example/tabs/utils/models/PresentHistory.kt
@@ -1,0 +1,8 @@
+package com.example.tabs.utils.models
+
+import java.util.Date
+
+data class PresentHistory(
+    val date: Date,
+    val gift: String
+)
